+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
-## [v1.0.1](https://github.com/Myoschen/muser/releases/tag/v1.0.1) (2022-12-15)
+<a name="1.0.1"></a>
 
-## [v1.0.0](https://github.com/Myoschen/muser/releases/tag/v1.0.0) (2022-11-25)
+## 1.0.1 (2022-12-22)
+
+### Changed
+
+- ⬆️ Upgrade deps [[19a77de](https://github.com/Myoschen/muser/commit/19a77de401906cdb3ed04172132083319ac5087b)]
+
+### Fixed
+
+- 🐛 Click elsewhere to close the popup [[00d86d5](https://github.com/Myoschen/muser/commit/00d86d52c13a22d62d750485db77095f6ccc15f6)]
+
+### Miscellaneous
+
+- fix(useClickAway): the event type of the handler [[d9e06c4](https://github.com/Myoschen/muser/commit/d9e06c45224d83de3fafc777d2c3c121155563a0)]
+- 📝 Update README.md and add CHANGELOG.md [[58e8c0d](https://github.com/Myoschen/muser/commit/58e8c0d0d60a47918f759c1ed536fb9f79e04447)]
+- docs: change the icon and preview images [[2486b1a](https://github.com/Myoschen/muser/commit/2486b1a3a29c52e89773ce0150f36442b9d31970)]
+- Merge branch &#x27;feature/watch-directory&#x27; into develop [[353d8d4](https://github.com/Myoschen/muser/commit/353d8d4dc01884541d84ad1a26fe9e496ab000c9)]
+- feat: auto reload audio list [[895bfc3](https://github.com/Myoschen/muser/commit/895bfc395e1522a520b5c583e4bf9039ac7cb6b2)]
+- feat: some changes to audio image and status bar [[13dd09a](https://github.com/Myoschen/muser/commit/13dd09aa3e25f29d67097897aebe917687a381c5)]
+- feat: custom radio style and new font [[5526734](https://github.com/Myoschen/muser/commit/5526734d4726042bcc5ca2df903ace4bb738c786)]
+- style: use destructuring to get props [[9b28b35](https://github.com/Myoschen/muser/commit/9b28b35acc74a14dc3c42594bedeb16c24299ad3)]
+- feat: use scada, kosugi maru, noto sans font family [[72cd82f](https://github.com/Myoschen/muser/commit/72cd82f47a03943d70ef663e6ce937082eefbf56)]
+- refactor: extract the relevant part of the styles [[3f42552](https://github.com/Myoschen/muser/commit/3f42552a721b199ba9d912c8b1c6f3f688907883)]
+- refactor: add error handle and channel enum [[5d698e3](https://github.com/Myoschen/muser/commit/5d698e3e3d713c50bf070b15761361447140e75d)]
+- Merge branch &#x27;feature/theme&#x27; into develop [[9eeabde](https://github.com/Myoschen/muser/commit/9eeabde846f8f6ebf2c5c259edb88c820d48e86c)]
+- fix: add class name to the document body [[ddacaa7](https://github.com/Myoschen/muser/commit/ddacaa7d555d47110fcc85d37e187c339d71d626)]
+- Merge branch &#x27;feature/theme&#x27; into develop [[bd31d0a](https://github.com/Myoschen/muser/commit/bd31d0a1fbb209278a7c86330e8fae9df571b6a6)]
+- fix: fix the scrollbar theme [[161fdcf](https://github.com/Myoschen/muser/commit/161fdcf6570d16d4adb49190606c9c97de25a17f)]
+- Merge branch &#x27;feature/theme&#x27; into develop [[e552cf5](https://github.com/Myoschen/muser/commit/e552cf503a6123d36fb1f354c8d6b90187d111d4)]
+- feat: implement the light and dark theme [[365b389](https://github.com/Myoschen/muser/commit/365b3890b080e483dd67f65d09067920bf6e50e2)]
+- fix: revise app module id, context menu lang [[9d9ebef](https://github.com/Myoschen/muser/commit/9d9ebefa69bfd99add075e5b6e4f80a2f7858f05)]
+- Merge branch &#x27;feature/setting&#x27; into develop [[c3e8c61](https://github.com/Myoschen/muser/commit/c3e8c613c536fbee07095bf22407fd22846976ed)]
+- feat: implement Setting modal [[9ffd0a8](https://github.com/Myoschen/muser/commit/9ffd0a8b52b47424041f269cbb147ba9284c7b4c)]
+- fix: invisible tray icon and ajv module error [[781b328](https://github.com/Myoschen/muser/commit/781b328d55eacf4dc6ccf674e8d5c8ba015efb30)]
+- Merge branch &#x27;feature/player&#x27; into develop [[393e473](https://github.com/Myoschen/muser/commit/393e473e5aa7821c1a5b2c3d44a2ac0d0d9b1250)]
+- feat: implement the audio player [[f19f661](https://github.com/Myoschen/muser/commit/f19f6618bef6fa4efde6c78e0e1452a8ca88bc9a)]
+- feat: implement some features [[cfec9e2](https://github.com/Myoschen/muser/commit/cfec9e2c4a7d2396fc42e4d55f1f570fbe01c00d)]
+- docs: edit something [[33535e7](https://github.com/Myoschen/muser/commit/33535e7c66510f03512efed8d5464e5e074050e5)]
+- initial commit [[3f389b8](https://github.com/Myoschen/muser/commit/3f389b8865c5df64bcb45851de5b66e6331b6824)]